+++ backend/src/main/java/org/ftclub/cabinet/admin/slack/controller/SlackController.java
@@ -1,9 +1,9 @@
-package org.ftclub.cabinet.slack.controller;
+package org.ftclub.cabinet.admin.slack.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.ftclub.cabinet.admin.slack.dto.SlackMessageDto;
+import org.ftclub.cabinet.admin.slack.service.SlackFacadeService;
 import org.ftclub.cabinet.log.Logging;
-import org.ftclub.cabinet.slack.dto.SlackMessageDto;
-import org.ftclub.cabinet.slack.service.SlackFacadeService;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @Logging
 @RestController
-@RequestMapping("/slack")
+@RequestMapping("/v5/admin/slack")
 @RequiredArgsConstructor
 public class SlackController {
