+++ backend/src/main/java/org/ftclub/cabinet/admin/slack/dto/SlackMessageDto.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.slack.dto;
+package org.ftclub.cabinet.admin.slack.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
